[airbyte-cdk] emit source recordCount as float instead of integer (#36560)

+++ airbyte-cdk/python/airbyte_cdk/entrypoint.py
@@ -161,26 +161,27 @@ class AirbyteEntrypoint(object):
         if self.source.check_config_against_spec:
             self.validate_connection(source_spec, config)
 
-        stream_message_counter: DefaultDict[HashableStreamDescriptor, int] = defaultdict(int)
+        # The Airbyte protocol dictates that counts be expressed as float/double to better protect against integer overflows
+        stream_message_counter: DefaultDict[HashableStreamDescriptor, float] = defaultdict(float)
         for message in self.source.read(self.logger, config, catalog, state):
             yield self.handle_record_counts(message, stream_message_counter)
         for message in self._emit_queued_messages(self.source):
             yield self.handle_record_counts(message, stream_message_counter)
 
     @staticmethod
-    def handle_record_counts(message: AirbyteMessage, stream_message_count: DefaultDict[HashableStreamDescriptor, int]) -> AirbyteMessage:
+    def handle_record_counts(message: AirbyteMessage, stream_message_count: DefaultDict[HashableStreamDescriptor, float]) -> AirbyteMessage:
         if message.type == Type.RECORD:
-            stream_message_count[message_utils.get_stream_descriptor(message)] += 1
+            stream_message_count[message_utils.get_stream_descriptor(message)] += 1.0
 
         elif message.type == Type.STATE:
             stream_descriptor = message_utils.get_stream_descriptor(message)
 
             # Set record count from the counter onto the state message
             message.state.sourceStats = message.state.sourceStats or AirbyteStateStats()
-            message.state.sourceStats.recordCount = stream_message_count.get(stream_descriptor, 0)
+            message.state.sourceStats.recordCount = stream_message_count.get(stream_descriptor, 0.0)
 
             # Reset the counter
-            stream_message_count[stream_descriptor] = 0
+            stream_message_count[stream_descriptor] = 0.0
         return message
 
     @staticmethod
